pyproject.toml
[tool.poetry]
name = "Finam_market_data_consumer"
version = "0.1.0"
description = "Simple rabbit_mq consumer for saving depth of market data to clickhouse"
authors = ["Alexander Evdokimov <83905079+snakerzr@users.noreply.github.com>"]
readme = "README.md"
#packages = [{include = "finam_market_data_consumer"}]

[tool.poetry.dependencies]
python = "^3.9"
pika = "^1.3.2"
clickhouse-driver = "^0.2.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
